Added some docs and reformatted code

+++ src/consts.h
@@ -6,8 +6,12 @@
 #define FML_CONSTS_H
 
 #include <cstddef>
+#include <cstdint>
 
-constexpr size_t CHUNK_SIZE = 50;
-constexpr size_t NUMBER_OF_THREADS = 50;
+/// The number of elements in a chunk
+constexpr uint8_t CHUNK_SIZE = 50;
+
+/// The number of supported threads
+constexpr uint8_t NUMBER_OF_THREADS = 50;
 
 #endif //FML_CONSTS_H
